--- src/Microsoft.AspNet.TelemetryCorrelation/.gitignore
@@ -1,15 +1,0 @@
-[Oo]bj/
-[Bb]in/
-TestResults/
-_ReSharper.*/
-/packages/
-artifacts/
-PublishProfiles/
-*.user
-*.suo
-*.cache
-*.sln.ide
-.vs
-.build/
-.testPublish/
-msbuild.*
