--- src/Microsoft.AspNet.TelemetryCorrelation/CODE-OF-CONDUCT.md
@@ -1,6 +1,0 @@
-# Code of Conduct
-
-This project has adopted the code of conduct defined by the Contributor Covenant
-to clarify expected behavior in our community.
-
-For more information, see the [.NET Foundation Code of Conduct](https://dotnetfoundation.org/code-of-conduct).
