--- src/Microsoft.AspNet.TelemetryCorrelation/test/Microsoft.AspNet.TelemetryCorrelation.Tests/Properties/AssemblyInfo.cs
@@ -1,18 +1,0 @@
-﻿// <copyright file="AssemblyInfo.cs" company="Microsoft">
-// Copyright (c) .NET Foundation. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-// </copyright>
-
-using System.Runtime.InteropServices;
-using Xunit;
-
-// Setting ComVisible to false makes the types in this assembly not visible
-// to COM components.  If you need to access a type in this assembly from
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("9fae5c43-f56c-4d87-a23c-6d2d57b4abed")]
-
-[assembly: CollectionBehavior(DisableTestParallelization = true)]
