--- src/Microsoft.AspNet.TelemetryCorrelation/tools/Common.props
@@ -1,64 +1,0 @@
-<Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-
- <PropertyGroup>
-   <LangVersion>7.3</LangVersion>
- </PropertyGroup>
-
-  <PropertyGroup>
-    <RepositoryRoot Condition="'$(RepositoryRoot)' == ''">$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory),Microsoft.AspNet.TelemetryCorrelation.sln))\</RepositoryRoot>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <Configuration Condition="'$(Configuration)' == ''">Release</Configuration>
-    <BinPath>$(RepositoryRoot)bin\</BinPath>
-    <ObjPath>$(RepositoryRoot)obj\</ObjPath>
-    <OutputPath>$(BinPath)$(Configuration)\</OutputPath>
-    <IntermediateOutputPath>$(ObjPath)$(Configuration)\$(MSBuildProjectName)\</IntermediateOutputPath>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <AdditionalFiles Include="$(MSBuildThisFileDirectory)/stylecop.json" />
-  </ItemGroup>
-
-  <!--VERSION -->
-    <PropertyGroup>
-      <BuildQuality Condition="'$(BuildQuality)' == ''">rtm</BuildQuality>
-      <VersionStartYear>2018</VersionStartYear>
-      <VersionMajor>1</VersionMajor>
-      <VersionMinor>0</VersionMinor>
-      <VersionRelease>9</VersionRelease>
-      
-      <VersionBuild Condition="'$(VersionBuild)' == '' OR '$(VersionBuild)' == '0'">$([MSBuild]::Add(1, $([MSBuild]::Subtract($([System.DateTime]::Now.Year), $(VersionStartYear)))))$([System.DateTime]::Now.ToString("MMdd"))</VersionBuild>
-      <VersionRevision Condition="'$(VersionRevision)' == ''">0</VersionRevision>
-      <VersionRelease Condition="'$(VersionRelease)' == ''">0</VersionRelease>
-    </PropertyGroup>
-
-    <PropertyGroup>
-      <VersionPrefix>$(VersionMajor).$(VersionMinor).$(VersionRelease)</VersionPrefix>
-      <VersionSuffix Condition="'$(BuildQuality)' != 'rtm'">$(BuildQuality)</VersionSuffix>
-      <FileVersion>$(VersionMajor).$(VersionMinor).$(VersionBuild).$(VersionRevision)</FileVersion>
-      <InformationalVersion>$(VersionMajor).$(VersionMinor).$(VersionRelease)-$(VersionBuild)</InformationalVersion>
-      <AssemblyVersion>$(VersionMajor).$(VersionMinor).$(VersionRelease).0</AssemblyVersion>
-    </PropertyGroup>
-
-    <PropertyGroup Label="Common Configuration">
-        <Configuration Condition="'$(Configuration)' == ''">Release</Configuration>
-        <OutputPath>$(RepositoryRoot)bin\$(Configuration)\</OutputPath>
-        <IntermediateOutputPath>$(RepositoryRoot)obj\$(Configuration)\$(MSBuildProjectName)\</IntermediateOutputPath>
-    </PropertyGroup>
-
-<!-- SIGNING -->
-    <ItemDefinitionGroup>
-      <FilesToSign>
-        <!-- Microsoft authenticode certificate for all file types -->
-        <Authenticode Condition="'%(FilesToSign.Authenticode)' == ''">Microsoft400</Authenticode>
-        <!-- Microsoft strong-name key for managed assemblies -->
-        <StrongName Condition="'$(TargetExt)' == '.winmdobj' OR '$(TargetExt)' == '.dll' OR '$(TargetExt)' == '.exe'">MsSharedLib72</StrongName>
-      </FilesToSign>
-    </ItemDefinitionGroup>
-
-    <ItemGroup>
-      <!-- Automatically sign binary generated by the project that imports Signing.props -->
-      <FilesToSign Include="$(TargetPath)" />
-    </ItemGroup>
-</Project>
