+++ src/Microsoft.AspNet.TelemetryCorrelation/web.config.install.xdt
