+++ src/Microsoft.AspNet.TelemetryCorrelation/web.config.uninstall.xdt
